[NO-JIRA] Lazy load konami
