+++ src/containers/common/Konami/ConfettiCanvasLoadable.js
@@ -1,0 +1,10 @@
+import Loadable from 'react-loadable';
+
+const Loading = () => null;
+
+const ConfettiCanvasLoadable = Loadable({
+  loader: () => import('./ConfettiCanvas.js'),
+  loading: Loading,
+});
+
+export default ConfettiCanvasLoadable;
